requirements.txt
altgraph==0.17.4
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.0
Jinja2==3.1.2
kiwisolver==1.4.4
macholib==1.16.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.0.9
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-dateutil==2.8.2
pytz==2023.3
shiboken6==6.6.1
six==1.16.0
sympy==1.12
torch==2.0.1
tqdm==4.65.0
trussty==0.1.1
typing_extensions==4.7.1
tzdata==2023.3
